Add full Direccion service implementation
Implemented DireccionService with CRUD operations, entity/DTO conversion, and user lookup. Updated Direccion model to include 'direccion' and 'altura' fields, changed id_usuario type in CreateDireccionDTO to Long, and added buscarEntidadPorId method to UsuarioService for user retrieval.

+++ src/main/java/org/example/topdeckapi/src/DTOs/CreateDTO/CreateDireccionDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CreateDireccionDTO {
     @NotBlank
-    private Integer id_usuario;
+    private Long id_usuario;
 
     @NotBlank
     private String ciudad;
